Add a cgi function to be used programatically for checking the block status of a url
I want to take a hosts file and turn it into a privoxy actions file blocklist.
I want it to be fast and I want it to drop host names that are already blocked.
I also want it to be fast, so that means not reading the disk for template files
or file change timestamps.

So make a copy of the cgi "show-url-info" function that
1) doesn't access the disk drive
2) doesn't do as much (less work = less time required)

Add cgi function show-url-final-info
- cgisimple.h
  declare extern cgi_show_url_final_info

- actions.h
  declare extern merge_single_actions

- actions.c
  merge_single_actions function copied from merge_current_action
   a bit faster since it doesn't process multi actions but I don't know
   what the downside is for not processing multi actions :(

- cgi.c
  - add entrypoint for "show-url-final-info"

- cgisimple.c
  - add a call to run_loader in cgi_show_status to make sure all config files are current
  - add cgi_show_url_final_info function to show final block state for a url

- jcc.c
  - don't check for config file changes if this is a csp request