bring back 'debug xxx' being interpreted as octal, hex or decimal
add a log message when the buffer-limit value is changed.

add a log message when the debug flags are changed.

-

The debug flags are binary bit flags.  It's easier for me to say
  debug 0x0400000
than convert 0x0400000 to decimal and make _that_ the value.
Which I'll have to convert back to binary whenever I want to
make sure the value is correct..

Which brings up another issue - it'd be nice to show which debug
flags are enabled in the show-status cgi page.